04 authentication

+++ src/restaurants/restaurants.controller.ts
@@ -2,12 +2,15 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   Param,
   Post,
   Put,
   Query,
+  Req,
   UploadedFiles,
+  UseGuards,
   UseInterceptors,
   UsePipes,
   ValidationPipe,
@@ -19,6 +22,11 @@ import { Restaurant } from './schemas/restaurant.schema';
 
 import { Query as ExpressQuery } from 'express-serve-static-core';
 import { FilesInterceptor } from '@nestjs/platform-express';
+import { AuthGuard } from '@nestjs/passport';
+import { CurrentUser } from '../auth/decorators/current-user.decorator';
+import { User } from '../auth/schemas/user.schema';
+import { RolesGuard } from '../auth/guards/roles.guard';
+import { Roles } from '../auth/decorators/roles.decorator';
 
 @Controller('restaurants')
 export class RestaurantsController {
@@ -30,11 +38,13 @@ export class RestaurantsController {
   }
 
   @Post()
-  // @UsePipes(ValidationPipe) -> change to global
+  @UseGuards(AuthGuard(), RolesGuard)
+  @Roles('admin', 'user')
   async createRestaurant(
     @Body() restaurant: CreateRestaurantDto,
+    @CurrentUser() user: User,
   ): Promise<Restaurant> {
-    return await this.restaurantsService.create(restaurant);
+    return await this.restaurantsService.create(restaurant, user);
   }
 
   @Get(':id')
@@ -43,19 +53,31 @@ export class RestaurantsController {
   }
 
   @Put(':id')
+  @UseGuards(AuthGuard())
   async updateRestaurant(
     @Param('id') id: string,
     @Body() restaurant: UpdateRestaurantDto,
+    @CurrentUser() user: User,
   ): Promise<Restaurant> {
-    await this.restaurantsService.findById(id);
+    const res = await this.restaurantsService.findById(id);
+
+    if (res.user._id.toString() !== user._id.toString())
+      throw new ForbiddenException('You can not update this restaurant');
+
     return this.restaurantsService.updateById(id, restaurant);
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuard())
   async deleteRestaurant(
     @Param('id') id: string,
+    @CurrentUser() user: User,
   ): Promise<{ deleted: Boolean }> {
     const restaurant = await this.restaurantsService.findById(id);
+
+    if (restaurant.user._id.toString() !== user._id.toString())
+      throw new ForbiddenException('You can not update this restaurant');
+
     const isDeleted = await this.restaurantsService.deleteImages(
       restaurant.images,
     );
